ex. 4.21

+++ part4/blogApp/controllers/blogs.js
@@ -49,8 +49,19 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  const userId = decodedToken.id;
+  if (blog.user.toString() === userId.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    response
+      .status(403)
+      .json({
+        error: "You do not have sufficient permissions to delete the resource",
+      });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
